+++ Gu.State.Tests/EqualByTests/TestCases.cs
@@ -6,6 +6,7 @@ namespace Gu.State.Tests.EqualByTests
     using System;
     using System.Collections.Generic;
     using System.Drawing;
+    using System.Linq;
     using NUnit.Framework;
 
     using static EqualByTypes;
@@ -51,6 +52,8 @@ namespace Gu.State.Tests.EqualByTests
 
             new TestCaseData(new[,] { { 1, 2 }, { 3, 4 }, { 5, 6 } }, new[,] { { 1, 2 }, { 3, 4 }, { 5, 6 } }),
 
+            new TestCaseData(Enumerable.Empty<int>().Select(x =>x * x), Enumerable.Empty<int>().Select(x => x * x)),
+            new TestCaseData("1,2".Split(',').Select(int.Parse), "1,2".Split(',').Select(int.Parse)),
             //new TestCaseData(
             //    new With<Point>(new Point(1, 2)),
             //    new With<Point>(new Point(1, 2))),
@@ -137,6 +140,9 @@ namespace Gu.State.Tests.EqualByTests
             new TestCaseData(new[,] { { 1, 2 }, { 3, 4 }, { 5, 6 } }, new[,] { { 1, 2 }, { -3, 4 }, { 5, 6 } }),
             new TestCaseData(new[,] { { 1, 2 }, { 3, 4 }, { 5, 6 } }, new[,] { { 1, 2 }, { 3, 4 }, { 5, -6 } }),
 
+            new TestCaseData(new[] { 1 }.Select(x => x * x), Enumerable.Empty<int>().Select(x => x * x)),
+            new TestCaseData("1,2".Split(',').Select(int.Parse), "1,-2".Split(',').Select(int.Parse)),
+            new TestCaseData("1,2".Split(',').Select(int.Parse), "1,2".Split(',').Select(int.Parse)),
             //new TestCaseData(
             //    new With<Point>(new Point(1, 2)),
             //    new With<Point>(new Point(1, 3))),
